// src/extension/models/constants.d.ts
export declare enum Constants {
    langCode = "dothttp-vscode",
    pythonPath = "dothttp.conf.pythonpath",
    dothttpPath = "dothttp.conf.path",
    experimental = "dothttp.conf.experimental",
    nocookie = "dothttp.conf.nocookie",
    history = "dothttp.conf.history",
    curl = "dothttp.conf.curl",
    showheaders = "dothttp.conf.showheaders",
    runConf = "dothttp.conf.runrecent",
    reUseOldTab = "dothttp.conf.run.reuseold",
    responseDirectory = "dothttp.conf.response.savedirectory",
    runFileCommand = "dothttp.command.run",
    genCurlForFileCommand = "dothttp.command.gencurl",
    toggleExperimentalCommand = "dothttp.command.toggle.experimental",
    toggleHistoryCommand = "dothttp.command.toggle.history",
    toggleNocookieCommand = "dothttp.command.toggle.nocookie",
    toggleHeadersCommand = "dothttp.command.toggle.showheaders",
    toggleReuseTabCommand = "dothttp.command.toggle.runrecent",
    toggleRunRecentCommand = "dothttp.command.toggle.reuse",
    importCommand = "dothttp.command.import.external",
    generateLangCommand = "dothttp.command.generatelang",
    envTreeView = "dothttpEnvView",
    dothttpHistory = "dothttpHistory",
    refreshEnvCommand = "dothttpEnvView.refresh",
    enableEnvCommand = "dothttpEnvView.enableenv",
    disableEnvCommand = "dothttpEnvView.disableenv",
    copyEnvValueCommand = "dothttpEnvView.copyPropertyValue",
    openEnvFileCommmand = "dothttpEnvView.opendothttpjson",
    disableAllEnvCommmand = "dothttpEnvView.disableAllEnv",
    enableEnvViewVar = "dothttpEnvViewEnabled",
    notebookscheme = "vscode-notebook-cell",
    propTreeView = "dothttpPropView",
    addPropCommand = "dothttpPropView.add",
    enablePropCommand = "dothttpPropView.enableproperty",
    disablePropCommand = "dothttpPropView.disableproperty",
    copyEnvPropCommand = "dothttpPropView.copyPropertyValue",
    updatePropCommand = "dothttpPropView.updateproperty",
    removePropCommand = "dothttpPropView.removeproperty",
    disableAllPropCommand = "dothttpPropView.disableAllProperties",
    propViewEnabled = "dothttpPropViewEnabled",
    NOTEBOOK_MIME_TYPE = "x-application/dotbook",
    extensionVersion = "0.0.15",
    dothttpNotebook = "dothttp-book",
    dothttpVersion = "dothttp.version",
    versionApi = "https://raw.githubusercontent.com/cedric05/dothttp-runner/VERSION/version.json"
}